refactor: update discount data types to int, remove unused add_discount_button, and enhance UI with new components

+++ lib/features/discount/presentation/cubit/discount_state.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 
 class DiscountItem {
   final int daysRequired;
-  final double discountPercentage;
+  final int discountPercentage;
 
   DiscountItem({required this.daysRequired, required this.discountPercentage});
 }
